+++ ci-droid-tasks-consumer-services/src/main/java/com/societegenerale/cidroid/tasks/consumer/services/model/github/UpdatedResource.java
@@ -72,7 +72,8 @@ public class UpdatedResource {
         UPDATE_KO_CANT_PROVIDE_CONTENT_ISSUE(false),
         UPDATE_KO_NO_REASON(false),
         UPDATE_KO_AUTHENTICATION_ISSUE(false),
-        UPDATE_KO_UNEXPECTED_EXCEPTION_DURING_PROCESSING(false);
+        UPDATE_KO_UNEXPECTED_EXCEPTION_DURING_PROCESSING(false),
+        UPDATE_KO_REPO_DOESNT_EXIST(false);
 
         private final boolean hasBeenUpdated;